+++ common/src/main/java/de/dennisthegamer/tradetracker/config/TradeTrackerConfig.java
@@ -2,7 +2,7 @@ package de.dennisthegamer.tradetracker.config;
 
 import com.google.gson.Gson;
 import com.google.gson.GsonBuilder;
-import net.fabricmc.loader.api.FabricLoader;
+import de.dennisthegamer.tradetracker.platform.Platforms;
 
 import java.io.File;
 import java.io.FileReader;
@@ -15,7 +15,7 @@ public class TradeTrackerConfig {
 
     private static final Gson GSON = new GsonBuilder().setPrettyPrinting().create();
     private static final File CONFIG_FILE = new File(
-            FabricLoader.getInstance().getConfigDir().toFile(),
+            Platforms.get().getConfigDir().toFile(),
             "tradetracker.json"
     );