+++ common/src/main/java/de/dennisthegamer/tradetracker/tracker/TradeSession.java
@@ -4,7 +4,7 @@ import de.dennisthegamer.tradetracker.TradeTrackerClient;
 import com.google.gson.Gson;
 import com.google.gson.GsonBuilder;
 import com.google.gson.reflect.TypeToken;
-import net.fabricmc.loader.api.FabricLoader;
+import de.dennisthegamer.tradetracker.platform.Platforms;
 
 import java.io.IOException;
 import java.lang.reflect.Type;
@@ -182,7 +182,7 @@ public class TradeSession {
     }
 
     private static final Gson GSON = new GsonBuilder().setPrettyPrinting().create();
-    private static final Path SESSION_FILE = FabricLoader.getInstance().getConfigDir()
+    private static final Path SESSION_FILE = Platforms.get().getConfigDir()
             .resolve("tradetracker_session.json");
 
     public void saveToDisk() {